+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+## [0.8.0](https://github.com/tari-project/tari_utilities/compare/v0.7...v0.8.0) (2024-10-10)
+
+
+### ⚠ BREAKING CHANGES
+
+* Changes the `ByteArray` trait.
+
+### Features
+
+* Changes the `ByteArray` trait.
+
 ### [0.7.0](https://github.com/tari-project/tari_utilities/compare/v0.6.1...v0.7.0) (2023-12-06)
 
 ### Features